feat: port database from Postgres to SQLite
- internal/db/db.go on modernc.org/sqlite, WAL pragmas, MaxOpenConns(1)
- goose dialect "sqlite3"; squash 5 migrations into 00001_initial.sql
- sqlc engine "sqlite"; queries moved postgres/ -> sqlite/; $N -> ?;
  ::text casts removed; strftime replacements for to_char/INTERVAL
- row_id SERIAL becomes INTEGER PRIMARY KEY AUTOINCREMENT; id flips to
  TEXT UNIQUE so v3 Radarr-compat row_id integer IDs round-trip and
  never-reuse semantics match Postgres SERIAL (VACUUM-stable)
- internal/api/v3/adapter.go: $1 -> ? in raw row_id<->UUID lookups
- backup.go ported from pg_dump/psql to VACUUM INTO + magic-byte
  validation; restore replaces the .db file with restart-required UX
- internal/core/dbutil: helpers operate on *string/*int64/*time.Time
  directly; IsUniqueViolation against *modernc/sqlite.Error
- DatabaseConfig collapses to {Path}; remove DefaultDBDriver and dead
  sqlite branch in load.go
- 3 Dockerfiles updated: PRISM_DATABASE_PATH=/config/prism.db, drop
  pgx-era ENV and password-file mounts
- int32->int64 / sql.NullString->*string / TIMESTAMPTZ->TEXT ripple
  across ~25 sqlc call sites in 14 packages

+++ internal/core/downloader/service.go
@@ -104,7 +104,7 @@ func (s *Service) Create(ctx context.Context, req CreateRequest) (Config, error)
 		Name:      req.Name,
 		Kind:      req.Kind,
 		Enabled:   req.Enabled,
-		Priority:  int32(priority),
+		Priority:  int64(priority),
 		Settings:  string(settings),
 		CreatedAt: now.Format(time.RFC3339),
 		UpdatedAt: now.Format(time.RFC3339),
@@ -172,7 +172,7 @@ func (s *Service) Update(ctx context.Context, id string, req UpdateRequest) (Con
 		Name:      req.Name,
 		Kind:      req.Kind,
 		Enabled:   req.Enabled,
-		Priority:  int32(priority),
+		Priority:  int64(priority),
 		Settings:  string(settings),
 		UpdatedAt: time.Now().UTC().Format(time.RFC3339),
 	})
